2 saveload

+++ src/prog2/model/AccesTerra.java
@@ -2,16 +2,16 @@ package prog2.model;
 
 public abstract class AccesTerra extends Acces {
     // atributs
-    private int longitud;
+    private float longitud;
 
     // constructor
-    public AccesTerra(String id, boolean estat, int longitud) {
+    public AccesTerra(String id, boolean estat, float longitud) {
         super(id, estat);
         this.longitud = longitud;
     }
 
     // setters i getters
-    public int getLongitud() {
+    public float getLongitud() {
         return longitud;
     }
