+++ src/prog2/model/CamiTerra.java
@@ -2,7 +2,7 @@ package prog2.model;
 
 public class CamiTerra extends AccesTerra {
     // constructor
-    public CamiTerra(String id, boolean estat, int longitud) {
+    public CamiTerra(String id, boolean estat, float longitud) {
         super(id, estat, longitud);
     }
